Fixing the infeasible ILPs and renew the printers

+++ TDA.py
@@ -9,7 +9,7 @@ class task (dict):
         dict.__setitem__(self, "deadline", float (deadline))
         dict.__setitem__(self, "exclusive-R", float (execlusiveR))
         dict.__setitem__(self, "resource", int (resource))
-        dict.__setitem__(self, "block", int (block))
+        dict.__setitem__(self, "block", float (block))
 
 def Workload_Contrained(T,C,t):
     return C*math.ceil((t)/T)
