+++ generator.py
@@ -16,7 +16,7 @@ class task (dict):
         dict.__setitem__(self, "deadline", float (deadline))
         dict.__setitem__(self, "exclusive-R", float (execlusiveR))
         dict.__setitem__(self, "resource", int (resource))
-        dict.__setitem__(self, "block", int (block))
+        dict.__setitem__(self, "block", float (block))
 
 def UUniFast(n,U_avg):
     global USet
